# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Feature Flag API
  description: "Feature flag management, evaluation, and audit APIs."
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Feature Flags
  description: "Feature flag management, evaluation, and audit APIs."
- name: Feature Flag Preview
  description: Preview proposed feature flag changes.
paths:
  /api/flags:
    post:
      tags:
      - Feature Flags
      summary: Create a feature flag
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeatureFlagRequest"
        required: true
      responses:
        "201":
          description: Feature flag created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FeatureFlagResponse"
        "400":
          description: Invalid request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Feature flag already exists
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/flags/{flagKey}/preview:
    post:
      tags:
      - Feature Flag Preview
      summary: Preview a feature flag change
      operationId: preview
      parameters:
      - name: flagKey
        in: path
        description: Feature flag key.
        required: true
        schema:
          type: string
          maxLength: 200
          minLength: 1
        example: checkout-redesign
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationPreviewRequest"
        required: true
      responses:
        "200":
          description: Preview diff
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EvaluationPreviewResponse"
        "400":
          description: Invalid request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Feature flag not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/evaluate:
    post:
      tags:
      - Feature Flags
      summary: Evaluate a feature flag
      operationId: evaluate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateFeatureFlagRequest"
        required: true
      responses:
        "200":
          description: Evaluation result
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EvaluateFeatureFlagResponse"
        "400":
          description: Invalid request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Feature flag not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/flags/{flagKey}:
    get:
      tags:
      - Feature Flags
      summary: Get a feature flag
      operationId: get
      parameters:
      - name: flagKey
        in: path
        description: Feature flag key.
        required: true
        schema:
          type: string
          maxLength: 200
          minLength: 1
        example: checkout-redesign
      responses:
        "200":
          description: Feature flag found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FeatureFlagResponse"
        "404":
          description: Feature flag not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    patch:
      tags:
      - Feature Flags
      summary: Update a feature flag
      operationId: update
      parameters:
      - name: flagKey
        in: path
        description: Feature flag key.
        required: true
        schema:
          type: string
          maxLength: 200
          minLength: 1
        example: checkout-redesign
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeatureFlagRequest"
        required: true
      responses:
        "200":
          description: Feature flag updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FeatureFlagResponse"
        "400":
          description: Invalid request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Feature flag not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/flags/{flagKey}/audit-events:
    get:
      tags:
      - Feature Flags
      summary: List audit events for a feature flag
      operationId: auditEvents
      parameters:
      - name: flagKey
        in: path
        description: Feature flag key.
        required: true
        schema:
          type: string
          maxLength: 200
          minLength: 1
        example: checkout-redesign
      responses:
        "200":
          description: Audit events ordered oldest first
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEventResponse"
        "404":
          description: Feature flag not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
components:
  schemas:
    CreateFeatureFlagRequest:
      type: object
      description: Request to create a feature flag.
      properties:
        flagKey:
          type: string
          description: Stable feature flag key.
          example: checkout-redesign
          maxLength: 200
          minLength: 1
        status:
          type: string
          description: Initial flag status.
          enum:
          - ENABLED
          - DISABLED
          example: ENABLED
        targetEnvironments:
          type: array
          description: Environments where the flag may be evaluated.
          example:
          - production
          - staging
          items:
            type: string
            enum:
            - development
            - staging
            - production
          minItems: 1
          uniqueItems: true
        killSwitchActive:
          type: boolean
          description: Whether the emergency kill switch is active.
          example: false
        tenantAllowlist:
          type: array
          description: Tenant IDs that are always allowed.
          example:
          - tenant-a
          items:
            type: string
            maxLength: 255
            minLength: 1
          maxItems: 1000
          minItems: 0
          uniqueItems: true
        rolloutPercentage:
          type: integer
          format: int32
          description: Percentage rollout from 0 to 100.
          example: 25
          maximum: 100
          minimum: 0
      required:
      - flagKey
      - killSwitchActive
      - rolloutPercentage
      - status
      - targetEnvironments
    FeatureFlagResponse:
      type: object
      description: Feature flag representation.
      properties:
        flagKey:
          type: string
          description: Stable feature flag key.
          example: checkout-redesign
        status:
          type: string
          description: Current flag status.
          enum:
          - ENABLED
          - DISABLED
          example: ENABLED
        targetEnvironments:
          type: array
          description: Target environments.
          example:
          - production
          - staging
          items:
            type: string
          uniqueItems: true
        killSwitchActive:
          type: boolean
          description: Whether the emergency kill switch is active.
          example: false
        tenantAllowlist:
          type: array
          description: Tenant IDs that are always allowed.
          example:
          - tenant-a
          items:
            type: string
          uniqueItems: true
        rolloutPercentage:
          type: integer
          format: int32
          description: Percentage rollout from 0 to 100.
          example: 25
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
    EvaluationPreviewContext:
      type: object
      description: Sample context to evaluate during preview.
      properties:
        environment:
          type: string
          description: Environment to evaluate.
          example: production
          maxLength: 255
          minLength: 1
        tenantId:
          type:
          - string
          - "null"
          description: Optional tenant ID to evaluate. Omit or send null when unknown.
          example: tenant-a
          maxLength: 255
          minLength: 0
          pattern: .*\S.*
        userId:
          type:
          - string
          - "null"
          description: Optional user ID to evaluate. Omit or send null when unknown.
          example: user-a
          maxLength: 255
          minLength: 0
          pattern: .*\S.*
      required:
      - environment
    EvaluationPreviewRequest:
      type: object
      description: Request to preview a proposed feature flag change.
      properties:
        proposedChange:
          $ref: "#/components/schemas/ProposedFeatureFlagChange"
          description: Partial change to apply in memory.
        sampleContexts:
          type: array
          description: Sample contexts to evaluate before and after the change.
          items:
            $ref: "#/components/schemas/EvaluationPreviewContext"
          maxItems: 100
          minItems: 0
      required:
      - proposedChange
      - sampleContexts
    ProposedFeatureFlagChange:
      type: object
      description: Partial feature flag change used only for preview.
      properties:
        status:
          type:
          - string
          - "null"
          description: Proposed flag status.
          enum:
          - ENABLED
          - DISABLED
          example: ENABLED
        targetEnvironments:
          type:
          - array
          - "null"
          description: Proposed target environments. Empty input clears target environments.
          example:
          - production
          - staging
          items:
            type: string
            enum:
            - development
            - staging
            - production
          uniqueItems: true
        killSwitchActive:
          type:
          - boolean
          - "null"
          description: Proposed kill switch state.
          example: false
        tenantAllowlist:
          type:
          - array
          - "null"
          description: Proposed tenant allowlist. Empty input clears the allowlist.
          example:
          - tenant-a
          items:
            type: string
            maxLength: 255
            minLength: 1
          maxItems: 1000
          minItems: 0
          uniqueItems: true
        rolloutPercentage:
          type:
          - integer
          - "null"
          format: int32
          description: Proposed percentage rollout from 0 to 100.
          example: 50
          maximum: 100
          minimum: 0
    EvaluationDiff:
      type: object
      description: Before and after evaluation result for one sample context.
      properties:
        environment:
          type: string
          description: Sample environment.
          example: production
        tenantId:
          type:
          - string
          - "null"
          description: Sample tenant ID.
          example: tenant-a
        userId:
          type:
          - string
          - "null"
          description: Sample user ID.
          example: user-a
        before:
          $ref: "#/components/schemas/EvaluationPreviewResult"
          description: Evaluation result for the current persisted flag.
        after:
          $ref: "#/components/schemas/EvaluationPreviewResult"
          description: Evaluation result for the proposed in-memory flag.
        changed:
          type: boolean
          description: Whether the enabled value changed.
          example: true
    EvaluationPreviewResponse:
      type: object
      description: Preview result for a proposed feature flag change.
      properties:
        flagKey:
          type: string
          description: Feature flag key.
          example: checkout-redesign
        diffs:
          type: array
          description: Per-sample evaluation diffs.
          items:
            $ref: "#/components/schemas/EvaluationDiff"
        summary:
          $ref: "#/components/schemas/EvaluationPreviewSummary"
          description: Aggregate counts for the preview.
    EvaluationPreviewResult:
      type: object
      description: Evaluation result captured inside a preview diff.
      properties:
        enabled:
          type: boolean
          description: Whether the flag is enabled.
          example: true
        reason:
          type: string
          description: Reason returned by the evaluator.
          enum:
          - FLAG_DISABLED
          - ENVIRONMENT_NOT_TARGETED
          - KILL_SWITCH_ACTIVE
          - TENANT_ALLOWLIST_MATCH
          - ROLLOUT_MATCH
          - ROLLOUT_MISS
          example: ROLLOUT_MATCH
        bucket:
          type:
          - integer
          - "null"
          format: int32
          description: "Rollout bucket, when bucket evaluation was used."
          example: 42
    EvaluationPreviewSummary:
      type: object
      description: Aggregate counts for a preview response.
      properties:
        sampleCount:
          type: integer
          format: int32
          description: Number of sample contexts evaluated.
          example: 3
        beforeEnabledCount:
          type: integer
          format: int32
          description: Samples enabled before the proposed change.
          example: 1
        beforeDisabledCount:
          type: integer
          format: int32
          description: Convenience count of samples disabled before the proposed change.
          example: 2
        enabledCount:
          type: integer
          format: int32
          description: Samples enabled after the proposed change.
          example: 2
        disabledCount:
          type: integer
          format: int32
          description: Convenience count of samples disabled after the proposed change.
          example: 1
        changedCount:
          type: integer
          format: int32
          description: Samples whose enabled value changed.
          example: 1
    EvaluateFeatureFlagRequest:
      type: object
      description: Request to evaluate a feature flag for a context.
      properties:
        flagKey:
          type: string
          description: Feature flag key to evaluate.
          example: checkout-redesign
          maxLength: 200
          minLength: 1
        environment:
          type: string
          description: Runtime environment.
          example: production
          maxLength: 255
          minLength: 1
        tenantId:
          type: string
          description: Optional tenant identifier.
          example: tenant-a
          maxLength: 255
          minLength: 0
        userId:
          type: string
          description: Optional user identifier.
          example: user-123
          maxLength: 255
          minLength: 0
      required:
      - environment
      - flagKey
    EvaluateFeatureFlagResponse:
      type: object
      description: Feature flag evaluation result.
      properties:
        flagKey:
          type: string
          description: Evaluated feature flag key.
          example: checkout-redesign
        enabled:
          type: boolean
          description: Whether the flag is enabled for the supplied context.
          example: true
        reason:
          type: string
          description: Reason for the evaluation result.
          enum:
          - FLAG_DISABLED
          - ENVIRONMENT_NOT_TARGETED
          - KILL_SWITCH_ACTIVE
          - TENANT_ALLOWLIST_MATCH
          - ROLLOUT_MATCH
          - ROLLOUT_MISS
          example: ROLLOUT_MATCH
        bucket:
          type: integer
          format: int32
          description: "Deterministic rollout bucket, when rollout logic was evaluated."
          example: 42
    UpdateFeatureFlagRequest:
      type: object
      description: Request to partially update a feature flag.
      properties:
        status:
          type: string
          description: Updated flag status.
          enum:
          - ENABLED
          - DISABLED
          example: DISABLED
        targetEnvironments:
          type: array
          description: Replacement target environments. Empty input clears target
            environments.
          example:
          - production
          items:
            type: string
            enum:
            - development
            - staging
            - production
          uniqueItems: true
        killSwitchActive:
          type: boolean
          description: Updated kill switch state.
          example: true
        tenantAllowlist:
          type: array
          description: Replacement tenant allowlist. Empty input clears the allowlist.
          example:
          - tenant-a
          - tenant-b
          items:
            type: string
            maxLength: 255
            minLength: 1
          maxItems: 1000
          minItems: 0
          uniqueItems: true
        rolloutPercentage:
          type: integer
          format: int32
          description: Updated percentage rollout from 0 to 100.
          example: 50
          maximum: 100
          minimum: 0
    AuditEventResponse:
      type: object
      description: Audit event emitted when a feature flag changes.
      properties:
        id:
          type: integer
          format: int64
          description: Audit event identifier.
          example: 42
        flagKey:
          type: string
          description: Feature flag key associated with the event.
          example: checkout-redesign
        eventType:
          type: string
          description: Type of audit event.
          enum:
          - FLAG_CREATED
          - FLAG_ENABLED
          - FLAG_DISABLED
          - ROLLOUT_PERCENTAGE_CHANGED
          - TARGET_ENVIRONMENTS_CHANGED
          - TENANT_ALLOWLIST_CHANGED
          - KILL_SWITCH_ENABLED
          - KILL_SWITCH_DISABLED
          example: ROLLOUT_PERCENTAGE_CHANGED
        details:
          description: Event-specific details. Shape depends on eventType.
        occurredAt:
          type: string
          format: date-time
          description: Time the event occurred.
          example: 2026-05-05T00:00:00Z