# service/src/main/kotlin/com/github/milez42/featureflags/preview/EvaluationPreviewController.kt
package com.github.milez42.featureflags.preview

import io.swagger.v3.oas.annotations.Operation
import io.swagger.v3.oas.annotations.Parameter
import io.swagger.v3.oas.annotations.media.Content
import io.swagger.v3.oas.annotations.media.Schema
import io.swagger.v3.oas.annotations.responses.ApiResponse
import io.swagger.v3.oas.annotations.responses.ApiResponses
import io.swagger.v3.oas.annotations.tags.Tag
import jakarta.validation.Valid
import jakarta.validation.constraints.NotBlank
import jakarta.validation.constraints.Size
import org.springframework.http.ProblemDetail
import org.springframework.validation.annotation.Validated
import org.springframework.web.bind.annotation.PathVariable
import org.springframework.web.bind.annotation.PostMapping
import org.springframework.web.bind.annotation.RequestBody
import org.springframework.web.bind.annotation.RequestMapping
import org.springframework.web.bind.annotation.RestController

@RestController
@RequestMapping("/api")
@Tag(name = "Feature Flag Preview", description = "Preview proposed feature flag changes.")
@Validated
class EvaluationPreviewController(private val service: EvaluationPreviewService) {
  @PostMapping("/flags/{flagKey}/preview")
  @Operation(summary = "Preview a feature flag change")
  @ApiResponses(
      value =
          [
              ApiResponse(
                  responseCode = "200",
                  description = "Preview diff",
                  content =
                      [Content(schema = Schema(implementation = EvaluationPreviewResponse::class))],
              ),
              ApiResponse(
                  responseCode = "400",
                  description = "Invalid request",
                  content = [Content(schema = Schema(implementation = ProblemDetail::class))],
              ),
              ApiResponse(
                  responseCode = "404",
                  description = "Feature flag not found",
                  content = [Content(schema = Schema(implementation = ProblemDetail::class))],
              ),
          ]
  )
  fun preview(
      @Parameter(
          description = "Feature flag key.",
          example = "checkout-redesign",
          schema = Schema(minLength = 1, maxLength = 200),
      )
      @PathVariable
      @NotBlank
      @Size(max = 200)
      flagKey: String,
      @Valid @RequestBody request: EvaluationPreviewRequest,
  ): EvaluationPreviewResponse = service.preview(flagKey, request)
}